container-api.sh: make container-api script work with zephyr-intel (#14)
Account for the zephyrproject/zephyr-intel vs zephyrproject/zephyr path
to repo checkout.

West needs to run against zephyr-intel dir and the quarantine list is
also located down that path. That's the only difference between zphyr
and zephyr-intel. So a simple check for a zephyrproject/zephyr-intel
directory tells us which flavor we have, and sets a ZEPHYR_DIR var to
use to toggle the path accordingly.

Signed-off-by: Graydon, Connor Jameson <connor.graydon@intel.com>