Remove usbip from Dockerfile
During my play with USBIP, I discovered, that there is a problem with my approach to using this package. 
I wanted to run the USB-Cilent directly in the container, but without success. There was a problem with accessing the device.
So my other approach is quite different. I decide to use the USBip client on the host machine, and pass the attached (remote) device to the container. It works.
I'd like to back my previous change.
https://github.com/intel-innersource/os.rtos.zephyr.devops.infrastructure.sdk-docker-intel/pull/92